+++ MercadoPagoSDK/MercadoPagoSDK/Services/Models/PXPointsAndDiscounts.swift
@@ -16,8 +16,9 @@ struct PXPointsAndDiscounts: Decodable {
     let topTextBox: PXText?
     let customOrder: Bool?
     let expenseSplit: PXExpenseSplit?
+    let paymentMethodsImages: [String: String]?
 
-    init(points: PXPoints?, discounts: PXDiscounts?, crossSelling: [PXCrossSellingItem]?, viewReceiptAction: PXRemoteAction?, topTextBox: PXText?, customOrder: Bool?, expenseSplit: PXExpenseSplit?) {
+    init(points: PXPoints?, discounts: PXDiscounts?, crossSelling: [PXCrossSellingItem]?, viewReceiptAction: PXRemoteAction?, topTextBox: PXText?, customOrder: Bool?, expenseSplit: PXExpenseSplit?, paymentMethodsImages: [String: String]?) {
         self.points = points
         self.discounts = discounts
         self.crossSelling = crossSelling
@@ -25,6 +26,7 @@ struct PXPointsAndDiscounts: Decodable {
         self.topTextBox = topTextBox
         self.customOrder = customOrder
         self.expenseSplit = expenseSplit
+        self.paymentMethodsImages = paymentMethodsImages
     }
 
     enum PointsAndDiscountsCodingKeys: String, CodingKey {
@@ -35,6 +37,7 @@ struct PXPointsAndDiscounts: Decodable {
         case topTextBox = "top_text_box"
         case customOrder = "custom_order"
         case expenseSplit = "expense_split"
+        case paymentMethodsImages = "payment_methods_images"
     }
 
     init(from decoder: Decoder) throws {
@@ -46,6 +49,7 @@ struct PXPointsAndDiscounts: Decodable {
         let topTextBox: PXText? = try container.decodeIfPresent(PXText.self, forKey: .topTextBox)
         let customOrder: Bool? = try container.decodeIfPresent(Bool.self, forKey: .customOrder)
         let expenseSplit: PXExpenseSplit? = try container.decodeIfPresent(PXExpenseSplit.self, forKey: .expenseSplit)
-        self.init(points: points, discounts: discounts, crossSelling: crossSelling, viewReceiptAction: viewReceiptAction, topTextBox: topTextBox, customOrder: customOrder, expenseSplit: expenseSplit)
+        let paymentMethodsImages: [String: String]? = try container.decodeIfPresent([String: String].self, forKey: .paymentMethodsImages)
+        self.init(points: points, discounts: discounts, crossSelling: crossSelling, viewReceiptAction: viewReceiptAction, topTextBox: topTextBox, customOrder: customOrder, expenseSplit: expenseSplit, paymentMethodsImages: paymentMethodsImages)
     }
 }