+++ MercadoPagoSDK/MercadoPagoSDK/UI/PXResult/New UI/PXNewResultViewModel.swift
@@ -60,6 +60,9 @@ protocol PXNewResultViewModelInterface {
 
     ////CUSTOM ORDER
     func getCustomOrder() -> Bool?
+    
+    ////PAYMENT METHOD IMAGES
+    func getPaymentMethodsImageURLs() -> [String: String]?
 
     //INSTRUCTIONS
     func hasInstructions() -> Bool